vector-add-buffers.cpp compilation fix
static_cast<int>( ) is added to fix the compilation error on Linux:
src/vector-add-buffers.cpp:147:26: error: non-constant-expression cannot be narrowed from type 'unsigned long' to 'int' in initializer list [-Wc++11-narrowing]
  int indices[]{0, 1, 2, (a.size() - 1)};

+++ DirectProgramming/DPC++/DenseLinearAlgebra/vector-add/src/vector-add-buffers.cpp
@@ -144,7 +144,7 @@ int main() {
     }
   }
 
-  int indices[]{0, 1, 2, (a.size() - 1)};
+  int indices[]{0, 1, 2, (static_cast<int>(a.size()) - 1)};
   constexpr size_t indices_size = sizeof(indices) / sizeof(int);
 
   // Print out the result of vector add.
